Desktop/AI-Research-Assistant-main/AI-Research-Assistant-main/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
crewai>=0.1.0
langchain-openai>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
